pyproject.toml
[tool.poetry]
name = "anx-interface"
version = "0.5.0"
description = "Implements anx_api"
authors = ["Aditya Shriwastava <aditya@flomobility.com>"]
readme = "README.md"
packages = [
  {include = "anx_interface"},
  {include = "anx_mock"},
  {include = "anx_logging"},
  {include = "anx_proto"}
]

[tool.poetry.dependencies]
python = "^3.8"
progress = "~1.6"
pyzmq = "~22.3.0"
opencv-python = "^4.7.0.68"
numpy = "^1.24.2"
pillow = "^9.4.0"
protobuf = "4.22.0"
psutil = "^5.9.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
